Refactor I/O operations into separate read/write modules
- Moved read functions to read.c and read.h
- Moved write functions to write.c and write.h
- Removed redundant functions from imm_abc.c and imm_dp.c
- Updated includes and function calls to use new modules
- Adjusted data types in imm_state_table.h for consistency

+++ imm_state_table.h
@@ -15,7 +15,7 @@ struct imm_state_table
   int nstates;
   uint16_t *ids;
   int16_t start_state_idx;
-  int end_state_idx;
+  int16_t end_state_idx;
   int8_t *span;
 
   // Debugging purpose
